move specifacation classes to sub fold
