+++ src/main/scala/org/apache/spark/streamdm/core/specification/ExampleSpecification.scala
@@ -14,7 +14,12 @@
  * limitations under the License.
  *
  */
-package org.apache.spark.streamdm.core
+package org.apache.spark.streamdm.core.specification
+
+//import org.apache.spark.streamdm.core.FeatureSpecification
+//import org.apache.spark.streamdm.core.InstanceSpecification
+//import org.apache.spark.streamdm.core.FeatureSpecification
+//import org.apache.spark.streamdm.core.InstanceSpecification
 
 /**
  * An ExampleSpecification contains information about the input and output
